+++ extensions/coolify/src/lib/components/open-in-coolify.tsx
@@ -1,8 +1,13 @@
-import { Action } from "@raycast/api";
+import { Action, Image, Keyboard } from "@raycast/api";
 import { generateCoolifyUrl } from "../utils";
 
-export default function OpenInCoolify(props: { url: string }) {
+export default function OpenInCoolify(props: { icon?: Image.ImageLike; title?: string; url: string }) {
   return (
-    <Action.OpenInBrowser icon="coolify.png" title="Open in Coolify" url={generateCoolifyUrl(props.url).toString()} />
+    <Action.OpenInBrowser
+      icon={props.icon || "coolify.png"}
+      title={props.title || "Open in Coolify"}
+      url={generateCoolifyUrl(props.url).toString()}
+      shortcut={Keyboard.Shortcut.Common.Open}
+    />
   );
 }